Move merge_dicts to common style utils
